chore: move title-bar into new dashboardsbar
