+++ src/components/DashboardsBarNew/TitleBar/ActionsBar.js
@@ -20,16 +20,16 @@ import { acSetDashboardStarred } from '../../../actions/dashboards.js'
 import { acClearItemFilters } from '../../../actions/itemFilters.js'
 import { acSetShowDescription } from '../../../actions/showDescription.js'
 import { apiPostShowDescription } from '../../../api/description.js'
-import ConfirmActionDialog from '../../../components/ConfirmActionDialog.js'
-import DropdownButton from '../../../components/DropdownButton/DropdownButton.js'
-import MenuItem from '../../../components/MenuItemWithTooltip.js'
+import ConfirmActionDialog from '../../ConfirmActionDialog.js'
+import DropdownButton from '../../DropdownButton/DropdownButton.js'
+import MenuItem from '../../MenuItemWithTooltip.js'
 import { useCacheableSection } from '../../../modules/useCacheableSection.js'
 import { orObject } from '../../../modules/util.js'
 import { sGetDashboardStarred } from '../../../reducers/dashboards.js'
 import { sGetNamedItemFilters } from '../../../reducers/itemFilters.js'
 import { sGetSelected } from '../../../reducers/selected.js'
 import { sGetShowDescription } from '../../../reducers/showDescription.js'
-import { ROUTE_START_PATH } from '../../start/index.js'
+import { ROUTE_START_PATH } from '../../../pages/start/index.js'
 import { apiStarDashboard } from './apiStarDashboard.js'
 import FilterSelector from './FilterSelector.js'
 import StarDashboardButton from './StarDashboardButton.js'
