+++ src/components/DashboardsBarNew/TitleBar/FilterDialog.js
@@ -31,8 +31,8 @@ import {
     acAddItemFilter,
     acRemoveItemFilter,
 } from '../../../actions/itemFilters.js'
-import { useSystemSettings } from '../../../components/SystemSettingsProvider.js'
-import { useUserSettings } from '../../../components/UserSettingsProvider.js'
+import { useSystemSettings } from '../../SystemSettingsProvider.js'
+import { useUserSettings } from '../../UserSettingsProvider.js'
 import { sGetItemFiltersRoot } from '../../../reducers/itemFilters.js'
 
 const FilterDialog = ({
